refactor: 인터페이스와 구현 클래스 사이에 추상 클래스 추가하여 중복 제거_1: FrameLast
